+++ src/interfaces/ipifyResponse.d.ts
@@ -1,0 +1,14 @@
+interface ipifyResponse {
+  ip: string;
+  location: {
+    city: string;
+    region: string;
+    postalCode: string;
+    timezone: string;
+    lat: number;
+    lng: number;
+  };
+  isp: string;
+}
+
+export default ipifyResponse;