+++ types/images.d.ts
@@ -1,0 +1,22 @@
+// Type definitions for images
+//TODO: Change the following with more specific types
+
+declare module "*.jpg" {
+  const value: any;
+  export default value;
+}
+
+declare module "*.png" {
+  const value: any;
+  export default value;
+}
+
+declare module "*.svg" {
+  const value: any;
+  export default value;
+}
+
+declare module "*.gif" {
+  const value: any;
+  export default value;
+}