+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 This project follows semantic versioning.
 
+## [0.2.0-alpha] - 2023-05-25
+
+Added
+
+- ipify api.
+- Learflet/React-leaflet map api.
+- Leaflet css.
+- Map, Main components.
+- Map style.
+
+Updated
+
+- Webpack files configuration.
+- App, Input, Dashboard components.
+
 ## [0.1.0-alpha] - 2023-04-24
 
 Added